needs to be unsigned

+++ rnd.c
@@ -3,8 +3,8 @@
 
 #include "rnd.h"
 
-long rnd_n(int n) {
-  long x, r;
+uint32_t rnd_n(uint32_t n) {
+  uint32_t x, r;
   do {
     x = rnd();
     r = x % n;
@@ -12,5 +12,5 @@ long rnd_n(int n) {
   return r;
 }
 
-void rnd_init(unsigned x) { srandom(x); }
-long rnd() { return random(); }
+void rnd_init(uint32_t x) { srandom(x); }
+uint32_t rnd() { return random(); }
